Start work :tada:
This is the template that we can use to get started and work on the report and summary. It can be used for the next seminar round.

Note that it may have been better to use the branch and PR system as that would allow us to do reviews. However we don't have much time so maybe save this more formal process when we will have longer to do writing.